refactor: Extract repository short name reader
Move functionality dealing with repository short names from FormBrowse
into RepositoryDescriptionProvider

No unit tests added due a number of concrete dependencies such as
file system and `Repositories.RepositoryHistory.Repositories`.
